Module cmd::cli::upgrade::v1_helper

👎Deprecated since 0.4.0: Used for migrate old version(v0.3) metadata

Modules

err 🔒 Deprecated

Macros

define_catalog_value 🔒 Deprecated

Structs

CATALOG_KEY_PATTERN 🔒 Deprecated

CatalogKeyDeprecated

CatalogValueDeprecated

SCHEMA_KEY_PATTERN 🔒 Deprecated

SchemaKeyDeprecated

SchemaValueDeprecated

TableGlobalValueDeprecated

Table global info contains necessary info for a datanode to create table regions, including table id, table meta(schema…), region id allocation across datanodes.

Constants

CATALOG_KEY_PREFIXDeprecated

NAME_PATTERN 🔒 Deprecated

The pattern of a valid catalog, schema or table name.

SCHEMA_KEY_PREFIXDeprecated