Module cmd::server

This module startups all the components of a TiKV server.

It is responsible for reading from configs, starting up the various server components, and handling errors (mostly by aborting and reporting to the user).

The entry point is run_tikv.

Components are often used to initialize other components, and/or must be explicitly stopped. We keep these components in the TiKVServer struct.

Structs

Servers

TiKVEngines

TiKVServer
A complete TiKV server.

Constants

RESERVED_OPEN_FDS

Traits

Stop
A small trait for components which can be trivially stopped. Lets us keep a list of these in TiKV, rather than storing each component individually.

Functions

check_system_config

get_lock_dir

pre_start
Various sanity-checks and logging before running a server.

run_tikv
Run a TiKV server. Returns when the server is shutdown by the user, in which case the server will be properly stopped.

try_lock_conflict_addr