Struct cmd::server::TiKVServer

struct TiKVServer<ER: RaftEngine> {
    config: TiKvConfig,
    cfg_controller: Option<ConfigController>,
    security_mgr: Arc<SecurityManager>,
    pd_client: Arc<RpcClient>,
    router: RaftRouter<RocksEngine, ER>,
    system: Option<RaftBatchSystem<RocksEngine, ER>>,
    resolver: PdStoreAddrResolver,
    state: Arc<Mutex<GlobalReplicationState>>,
    store_path: PathBuf,
    encryption_key_manager: Option<Arc<DataKeyManager>>,
    engines: Option<TiKVEngines<ER>>,
    servers: Option<Servers<ER>>,
    region_info_accessor: RegionInfoAccessor,
    coprocessor_host: Option<CoprocessorHost<RocksEngine>>,
    to_stop: Vec<Box<dyn Stop>>,
    lock_files: Vec<File>,
    concurrency_manager: ConcurrencyManager,
}

A complete TiKV server.

Fields
config: TiKvConfigcfg_controller: Option<ConfigController>security_mgr: Arc<SecurityManager>pd_client: Arc<RpcClient>router: RaftRouter<RocksEngine, ER>system: Option<RaftBatchSystem<RocksEngine, ER>>resolver: PdStoreAddrResolverstate: Arc<Mutex<GlobalReplicationState>>store_path: PathBufencryption_key_manager: Option<Arc<DataKeyManager>>engines: Option<TiKVEngines<ER>>servers: Option<Servers<ER>>region_info_accessor: RegionInfoAccessorcoprocessor_host: Option<CoprocessorHost<RocksEngine>>to_stop: Vec<Box<dyn Stop>>lock_files: Vec<File>concurrency_manager: ConcurrencyManager
Implementations

impl<ER: RaftEngine> TiKVServer<ER>

fn init(config: TiKvConfig) -> TiKVServer<ER>

fn init_config(config: TiKvConfig) -> ConfigController

Initialize and check the config

Warnings are logged and fatal errors exist.

Fatal errors

If dynamic config feature is enabled and failed to register config to PD

If some critical configs (like data dir) are differrent from last run

If the config can't pass validate()

If the max open file descriptor limit is not high enough to support the main database and the raft database.

fn connect_to_pd_cluster(
config: &mut TiKvConfig,
security_mgr: Arc<SecurityManager>
) -> Arc<RpcClient>

fn check_conflict_addr(&mut self)

fn init_fs(&mut self)

fn init_yatp(&self)

fn init_encryption(&mut self)

fn create_raftstore_compaction_listener(&self) -> CompactionListener

fn init_engines(&mut self, engines: Engines<RocksEngine, ER>)

fn init_gc_worker(
&mut self
) -> GcWorker<RaftKv<ServerRaftStoreRouter<RocksEngine, ER>>, RaftRouter<RocksEngine, ER>>

fn init_servers(
&mut self,
gc_worker: &GcWorker<RaftKv<ServerRaftStoreRouter<RocksEngine, ER>>, RaftRouter<RocksEngine, ER>>
) -> Arc<ServerConfig>

fn register_services(&mut self)

fn init_metrics_flusher(&mut self)

fn run_server(&mut self, server_config: Arc<ServerConfig>)

fn run_status_server(&mut self)

fn stop(self)

impl TiKVServer<RocksEngine>

fn init_raw_engines(&mut self) -> Engines<RocksEngine, RocksEngine>

impl TiKVServer<RaftLogEngine>

fn init_raw_engines(&mut self) -> Engines<RocksEngine, RaftLogEngine>