Trait cmd::server::Stop

trait Stop {
    fn stop(self: Box<Self>);
}

A small trait for components which can be trivially stopped. Lets us keep a list of these in TiKV, rather than storing each component individually.

Required methods

fn stop(self: Box<Self>)

Loading content...