cmd

Module standalone

Structs

Command

DefaultNodeManagerCreator

DefaultProcedureExecutorCreator

DefaultTableIdAllocatorCreator 🔒

Instance

InstanceCreator

InstanceCreator is used for grouping various component creators for building the Standalone instance, suitable for customizing how the instance can be built.

InstanceCreatorResult

InstanceCreatorResult is expected to be used paired with InstanceCreator. It stores the created and other important components for further reusing.

StartCommand

Enums

SubCommand 🔒

Constants

APP_NAME

Traits

NodeManagerCreator

ProcedureExecutorCreator

TableIdAllocatorCreator