Crate cmd

Modules

cli

datanode

error

flownode

frontend

metasrv

options

standalone

Structs

APP_VERSION 🔒

CPU_LIMIT 🔒

MEMORY_LIMIT 🔒

Traits

App

Functions

create_resource_limit_metrics

log_env_flags 🔒

log_versions

Log the versions of the application, and the arguments passed to the cli.

start_wait_for_close_signal 🔒

wait for the close signal, for unix platform it’s SIGINT or SIGTERM